moved core to components root
