+++ src/js/components/MenuBar.js
