+++ src/js/components/Settings.js
