+++ src/js/components/SideBar/StatusBar.js
