+++ src/js/components/Style.js
