+++ src/js/components/SwitchCheck.js
