+++ src/js/components/ToolDescription.js
