+++ src/js/components/groupMenu/Group.js
