+++ src/js/components/groupMenu/GroupItem.js
