+++ src/js/components/groupMenu/Style.js
