+++ src/js/components/licenses/Licenses.js
