+++ src/js/components/licenses/ThirdParty.js
