+++ src/js/components/licenses/newLicenses.json
