+++ src/js/components/login/pages/CreativeCommonsPage.js
