+++ src/js/components/login/pages/StatementOfFaithPage.js
