+++ src/js/components/login/pages/TermsAndConditionsPage.js
