+++ tests/e.ProjectManifest.js
@@ -1,5 +1,5 @@
 import {assert} from 'chai';
-import ManifestGenerator from '../src/js/components/core/createProject/ProjectManifest.js';
+import ManifestGenerator from '../src/js/components/createProject/ProjectManifest.js';
 
 const sampleData = {
   repo: 'https://github.com/unfoldingWord-dev/translationCore/',
