+++ tests/g.GitApi.js
@@ -4,7 +4,7 @@
  ******************************************************************************/
 
 const assert = require('chai').assert;
-const GitApi = require('../src/js/components/core/GitApi.js');
+const GitApi = require('../src/js/components/GitApi.js');
 const CoreStore = require('../src/js/stores/CoreStore.js');
 const fs = require('fs-extra');
